PPHA-705: Move metrics commands into questions domain
